+++ inst/shinyApps/LogViewer/ui.R
@@ -5,12 +5,19 @@ shinyUI(
   fluidPage(
     titlePanel(paste("Log File Viewer -", logFileName)),
     fluidRow(
-      column(1,
+      column(2,
+             actionButton("firstButton", label = "|<", title = "First chunk"),
+             actionButton("backButton", label = "<", title = "Previous chunk"),
+             actionButton("forwardButton", label = ">", title = "Next chunk"),
+             actionButton("lastButton", label = ">|", title = "Last chunk"),
+             uiOutput("progressUi"),
              selectInput("level", label = "Level", choices = levels, selected = "INFO"),
-             selectInput("thread", label = "Thread", choices = threads),
-             selectInput("package", label = "Package", choices = packages)),
-      column(11,
-             dataTableOutput("logTable"))
+             selectInput("thread", label = "Thread", choices = c()),
+             selectInput("package", label = "Package", choices = c())
+             ),
+      column(10,
+             dataTableOutput("logTable"),
+             uiOutput("detailsUi"))
     )
   )
 )