+++ tests/testthat.R
@@ -1,2 +1,3 @@
 library(testthat)
+library(ParallelLogger)
 test_check("ParallelLogger")