+++ tests/testthat/test-args.R
@@ -1,0 +1,7 @@
+library("testthat")
+
+test_that("Creation of args functions", {
+  argsFun <- createArgFunction("createLogger")
+  expect_true(any(grepl("#' @param", argsFun)))
+  expect_true(any(grepl("createCreateLoggerArgs <- function", argsFun)))
+})